requirements.txt
PyAutoGUI~=0.9.54
pyperclip~=1.9.0
selenium~=4.27.1
webdriver-manager~=4.0.2
beautifulsoup4~=4.12.3
mysql-connector-python~=9.1.0
python-dotenv~=1.0.1
PyYAML~=6.0.2